+++ src/pkgcore/sync/bzr.py
@@ -1,8 +1,7 @@
 __all__ = ("bzr_syncer",)
 
 import os
-
-from snakeoil.process.spawn import spawn_get_output
+import subprocess
 
 from . import base
 
@@ -17,11 +16,13 @@ class bzr_syncer(base.VcsSyncer):
         bzr_path = os.path.join(path, ".bzr")
         if cls.disabled or not os.path.isdir(bzr_path):
             return None
-        code, data = spawn_get_output([cls.binary, "info", path])
-        if code != 0:
+        ret = subprocess.run(
+            [cls.binary, "info", path], capture_output=True, text=True, check=False
+        )
+        if ret.returncode != 0:
             # should alert the user somehow
             return None
-        for line in data:
+        for line in ret.stdout.splitlines():
             line = line.strip().split(":", 1)
             if len(line) != 2:
                 continue